// src/app/api/model/getCodeValuesDataResponse.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * GetCodeValuesDataResponse
 */
export interface GetCodeValuesDataResponse { 
    active?: boolean;
    description?: string;
    id?: number;
    mandatory?: boolean;
    name?: string;
    position?: number;
}